requirements.txt
Django==4.2.1
djangorestframework==3.14.0
drf-yasg==1.21.5
django-redis==5.2.0
redis==4.5.5
